fix: Apple Health sync build errors + barcode scanner rebuild
HealthManager (Services):
- Fixed HealthKit API calls for iOS 18 (deprecated signatures)
- Changed HealthImportLog to struct (not @Model)
- Fixed date formatting toString() -> String(...)
- Proper HKObjectType query options

HealthSyncView (Health):
- All HealthManager calls updated to shared singleton
- Fixed checkAuthorization -> requestAuthorization
- Fixed sleep hours: Double type for fractional hours (e.g. 7.5)
- Removed invalid CoreData insert for non-model struct

DailyLog (Models):
- Changed sleepHours from Int? to Double? for accuracy
- Fixes type mismatch in Dashboard

DashboardView:
- Fixed loggedSleepHours bindings (Int/Double conversions)
- Sleep logging sheet: hours now Int, totalMinutes calculation fixed
- Added keyboard dismissal button on chat (previous fix)

BarcodeScannerView (Camera):
- Complete rebuild (file was corrupted, 116 bytes)
- Manual barcode entry with keypad
- 20 Indian food products in local fallback DB
- Nutrition facts sheet with servings adjuster
- One-tap logging to DailyLog

+++ App/Models/DailyLog.swift
@@ -15,7 +15,7 @@ final class DailyLog {
     
     // Health metrics
     var steps: Int?
-    var sleepHours: Int?
+    var sleepHours: Double?
     var sleepMinutes: Int?
     var heartRate: Int?
     var waterML: Int?
@@ -32,7 +32,7 @@ final class DailyLog {
         goalCompleted: Int? = nil,
         imageUri: String? = nil,
         steps: Int? = nil,
-        sleepHours: Int? = nil,
+        sleepHours: Double? = nil,
         sleepMinutes: Int? = nil,
         heartRate: Int? = nil,
         waterML: Int? = nil,
